+++ components/MenuItem/index.js
@@ -4,38 +4,43 @@ import styled from 'styled-components';
 import {withTheme} from 'styled-components';
 
 import IcArrow from '../../icons/IcArrowForward';
+import Block from '../Block';
 
-const Container = styled.div`
-  ${props => props.theme.parseMargin(props.theme, props)};
-`;
+const parseWidth = ({px, theme}) => {
+  if (!px) return '';
+  if (typeof px === 'string') return `width: calc(100% - (${px} * 2))`;
+  else return `width: calc(100% - ${px * theme.factors.padding * 2}px)`;
+};
+
 const InnerContainer = styled.a`
   display: flex;
   align-items: center;
   justify-content: space-between;
   border-bottom: 1px solid rgba(0,0,0,0.1);
-  color: ${props => props.theme.colors.activeColor};
+  color: ${props => props.$color ? props.theme.parseColor(props, props.theme, '$color') : props.theme.colors.activeColor};
   text-decoration: unset;
   cursor: pointer;
   svg {
     width: 24px;
     height: 24px;
-    fill: ${props => props.theme.colors.activeColor};
+    fill: ${props => props.$color ? props.theme.parseColor(props, props.theme, '$color') : props.theme.colors.activeColor};
   }
   &:hover {
     background-color: ${props => props.theme.colors.neutral10};
     opacity: 0.8;
   }
-  ${props => props.theme.parsePadding(props.theme, props)};
+  ${props => props.theme.parsePadding(props, props.theme)};
+  ${props => parseWidth(props)};
 `;
 
-const MenuItem = ({children, path, mb, px, py}) => {
+const MenuItem = ({children, path, mb, px, py, color}) => {
   return (
-    <Container mb={mb}>
-      <InnerContainer href={path} px={px} py={py}>
+    <Block width='100%' mb={mb}>
+      <InnerContainer $color={color} href={path} px={px} py={py}>
         {children ? children : <span>Content here</span>}
         <IcArrow />
       </InnerContainer>
-    </Container>
+    </Block>
   );
 };
 
@@ -64,7 +69,11 @@ MenuItem.propTypes = {
   /**
    * Realiza a navegação do usuário através do clique no componente
    */
-  path: PropTypes.string.isRequired
+  path: PropTypes.string.isRequired,
+  /**
+   * Permite customizar a cor do texto e ícone
+   */
+  color: PropTypes.string
 };
 
 export default withTheme(MenuItem);